backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
scikit-learn==1.7.2
numpy==2.3.3
joblib==1.5.2
nltk==3.9.1
PyJWT==2.10.1
requests==2.32.3
